increase timeout

+++ components/webserver/interface/webserver.h
@@ -25,7 +25,7 @@ typedef struct {
     uint8_t ap_record_id;   //< ID of chosen AP. It can be used to access ap_records array from wifi_controller - ap_scanner
     uint8_t type;           //< Chosen type of attack
     uint8_t method;         //< Chosen method of attack
-    uint8_t timeout;        //< Attack timeout in seconds
+    uint16_t timeout;        //< Attack timeout in seconds
 } attack_request_t;
 
 /**
